requirements.txt
numpy>=1.26
pandas>=2.2
matplotlib>=3.8
pytest>=8.0
PyYAML>=6.0
